requirements.txt
viam-sdk
dht11
